Fix sxg_sig_test to allow variable-size sig. (#77)
The ECDSA signature is a DER-encoded pair of integers, which results in
variable length. Change the assertion to remove the signature rather than
filling with % chars, so that it's resilient to this.

In addition, add an assertion that the signature is correct, by hard-coding an
expected signed message and then making all the necessary OpenSSL calls.